nucliadb_utils/requirements.txt
pydantic==1.9.0
nats-py[nkeys]==2.0.0
aioredis==2.0.1
orjson==3.6.7
lru-dict==1.1.7
aiohttp==3.7.4
backoff==1.11.1
oauth2client==4.1.3
aiobotocore==2.1.1
types-aiofiles==0.8.3
google-api-python-client==2.37.0
types-aiofiles==0.8.3
types-requests==2.27.7
starlette
prometheus_client
google-auth==2.4.1
mmh3==3.0.0
